move text rendering logic to prismic helpers
